requirements.txt
pandas==1.5.3
tqdm==4.65.0
numpy==1.24.2
torch==1.13.0
scikit_learn==1.2.2
pytorch_lightning==1.5.10
hydra_core==1.3.2
mlflow==2.1.1
pre-commit==3.5.0
